- Improve accuracy of field-of-view calculation - Disable histo equalisation checkbox if OpenCV version < 2.6 (openCV hist eq not available in old versions) - Make raycaster accept integer pixel numbers without complaining. - Add undistort_image method in CalibResults for un-distorting images from a camera (but I much prefer building the distortion properly in models for camera data, rather than trying to remove distortion from the data).

+++ calcam/raytrace.py
@@ -111,9 +111,9 @@ class RayCaster:
         Results = RayData()
         Results.ResultType = 'PixelRayCast'
         Results.fullchip = fullchip
-        Results.x = np.copy(x)
+        Results.x = np.copy(x).astype('float')
         Results.x[valid_mask == 0] = 0
-        Results.y = np.copy(y)
+        Results.y = np.copy(y).astype('float')
         Results.y[valid_mask == 0] = 0
         Results.transform = self.fitresults.transform
